// P4/src/types/Task.ts
import { ObjectId } from "mongodb";

export type TaskStatus = "PENDING" | "IN_PROGRESS" | "COMPLETED";
export type TaskPriority = "LOW" | "MEDIUM" | "HIGH";

export type Task = {
    _id?: ObjectId;
    title: string;
    projectId: ObjectId; 
    assignedTo?: ObjectId; 
    status: TaskStatus; 
    priority: TaskPriority;
    dueDate?: Date;
};